protecting resolvers that need authentication

+++ src/graphql/resources/comment/comment.resolvers.ts
@@ -3,6 +3,9 @@ import { Dbconnection } from "../../../Interfaces/DbConnectionInterface";
 import { Transaction } from "sequelize";
 import { CommentInstance } from "../../../Models/CommentModel";
 import { handleError } from "../../../utils/utils";
+import { compose } from "../../composable/composable.resolver";
+import { authResolvers } from "../../composable/auth.resolver";
+import { AuthUser } from "../../../Interfaces/AuthUserInterface";
 
 export const commentResolvers = {
     Comment: {
@@ -26,32 +29,42 @@ export const commentResolvers = {
     },
 
     Mutation: {
-        createComment: (parent, {input}, { db }: {db: Dbconnection}, info: GraphQLResolveInfo) => {
+        createComment: compose(...authResolvers)((parent, {input}, { db, authUser }: {db: Dbconnection, authUser: AuthUser}, info: GraphQLResolveInfo) => {
+            input.user = authUser.id
             return db.sequelize.transaction((t: Transaction) => {
                 return db.Comment.create(input, {transaction: t});
             }).catch(handleError);
-        },
-        updateComment: (parent, {id, input}, { db }: {db: Dbconnection}, info: GraphQLResolveInfo) => {
-            id = parseInt(id)
+        }),
+
+        updateComment: compose(...authResolvers)((parent, {id, input}, { db, authUser }: {db: Dbconnection, authUser: AuthUser}, info: GraphQLResolveInfo) => {
+            id = parseInt(id);
             return db.sequelize.transaction((t: Transaction) => {
                 return db.Comment.findById(id).then((comment: CommentInstance) => {
                     if(!comment) {
                         throw new Error (`Comment with id ${id} not found!`);
                     }
+                    if(comment.get('user') != authUser.id) {
+                        throw new Error ('Cannot edit comment from another users!')
+                    }
+                    input.user = authUser.id;
                     return comment.update(input, {transaction: t});
                 })
             }).catch(handleError);
-        },
-        deleteComment: (parent, {id}, { db }: {db: Dbconnection}, info: GraphQLResolveInfo) => {
-            id = parseInt(id)
+        }),
+
+        deleteComment: compose(...authResolvers)((parent, {id}, { db, authUser }: {db: Dbconnection, authUser: AuthUser}, info: GraphQLResolveInfo) => {
+            id = parseInt(id);
             return db.sequelize.transaction((t: Transaction) => {
                 return db.Comment.findById(id).then((comment: CommentInstance) => {
                     if(!comment) {
                         throw new Error (`Comment with id ${id} not found!`);
+                    }
+                    if(comment.get('user') != authUser.id) {
+                        throw new Error ('Cannot delete comment from another users!')
                     }
                     return comment.destroy({transaction: t}).then(comment => !!comment);
                 })
             }).catch(handleError);
-        },
+        }),
     }
 }
